Get rid of position_map _extras.c file.
Entire-Checkpoint: 01M23ZEHJ8X2NGPV2V1BHCS7PK
